requirements.txt
asgiref==3.7.2
certifi==2024.2.2
Django==5.0.2
PyJWT==2.8.0
sentry-sdk==1.40.3
sqlparse==0.4.4
tabulate==0.9.0
typing_extensions==4.9.0
urllib3==2.2.0
